BugzID: 92851 - Revise method of detecting new API
Currently the method for checking the presence of the new Reporting
feature in VectorCAST relies on the presence of a single file in the
distrobution. However this file will not be present in the future.
Instead an import should be attempted. If the import name ever
changes in the future, it would probably be worth it to just check
the VectorCAST version from either clicast or look in the tool
version text file.